Refactor asset tree structure; remove obsolete controller and payload files, and reorganize tree-related models and states for improved clarity and maintainability.
