+++ lib/ui/features/assets/widgets/tree/assets_tree_node_model.dart
