+++ lib/ui/features/assets/widgets/tree/assets_tree_view.dart
@@ -2,8 +2,8 @@ import 'dart:isolate';
 
 import 'package:flutter/material.dart';
 
-import '../../../../core/domain/company_asset/company_asset.dart';
-import '../../../../core/domain/company_location/company_location.dart';
+import '../../../../../core/domain/company_asset/company_asset.dart';
+import '../../../../../core/domain/company_location/company_location.dart';
 import 'assets_tree_isolate.dart';
 import 'assets_tree_state.dart';
